refactor(kfx): keep Python service bootstrap in host boundary
